Move quantiles.hpp files to legacy
