+++ cpp/benchmarks/quantiles/group_quantiles_benchmark.cu
@@ -16,7 +16,7 @@
 
 #include <tests/utilities/legacy/column_wrapper.cuh>
 
-#include <cudf/quantiles.hpp>
+#include <cudf/legacy/quantiles.hpp>
 #include <random>
 
 #include <benchmark/benchmark.h>
