+++ cpp/include/cudf/legacy/io_types.hpp
@@ -21,7 +21,7 @@
 #include <memory>
 #include <utility>
 
-#include <cudf/cudf.h>
+// #include <cudf/cudf.h>
 #include <cudf/types.hpp>
 #include <cudf/legacy/table.hpp>
 #include <cudf/legacy/io_types.h>
