+++ cpp/src/quantiles/legacy/group_quantiles.cu
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-#include "group_quantiles.hpp"
+#include <quantiles/legacy/group_quantiles.hpp>
 
 #include <utilities/cuda_utils.hpp>
-#include <quantiles/quantiles_util.hpp>
+#include <quantiles/legacy/quantiles_util.hpp>
 #include <groupby/sort/legacy/sort_helper.hpp>
 #include <cudf/utilities/legacy/type_dispatcher.hpp>
