+++ cpp/src/quantiles/legacy/quantiles.cu
@@ -17,7 +17,7 @@
 //Quantile (percentile) functionality
 
 #include <cudf/cudf.h>
-#include <quantiles/quantiles_util.hpp>
+#include <quantiles/legacy/quantiles_util.hpp>
 #include <utilities/cudf_utils.h>
 #include <cudf/utilities/error.hpp>
 #include <cudf/utilities/legacy/type_dispatcher.hpp>
